requirements.txt
python-dotenv==1.0.0
streamlit==1.28.1
unstructured==0.10.30
tiktoken==0.5.1
langchain==0.1.0
langchain-chroma==0.0.1
langchain-groq==0.0.1
langchain-community==0.0.10
protobuf==4.25.1
langchain-huggingface==0.0.6
sentence-transformers==2.2.2
chromadb==0.4.18